+++ FrontEnd/src/App.js
@@ -3,7 +3,6 @@ import { BrowserRouter, Route,Routes } from "react-router-dom";
 import OpenRoom from './components/OpenRoom.js';
 import EnterAddress from './components/EnterAddress.js';
 import Main from './components/Main.js';
-import Menu from './components/Menu';
 import Historypage from './components/Historypage.js';
 
 
@@ -15,7 +14,6 @@ function App() {
       <BrowserRouter>
       <Routes>
         <Route exact path="/" element={<Main />}/>
-        <Route path="/menu" element={<Menu />}/>
         <Route path="/history" element={<Historypage />}/>
         <Route path="/openroom" element={<OpenRoom />}/>
         <Route path="/enteraddress" element={<EnterAddress />}/>
