+++ FrontEnd/src/components/Historypage.module.scss
@@ -11,10 +11,12 @@
     justify-content:center;
     text-align: left;
     color: black;
+    
 }
 .row{
     position: relative;
     top: 80px;
+    z-index: -1;
 }
 .col{
     width: 85vw;
