--- FrontEnd/src/components/Menu/Menu.module.scss
@@ -1,56 +1,0 @@
-@import url('https://cdn.rawgit.com/moonspam/NanumSquare/master/nanumsquare.css');
-@mixin ani() {
-    text-decoration: none;
-    color: #F2A341;
-    opacity: 0.5;
-    transition: opacity 0.2s ease;
-}
-
-.overlay{
-    font-family: 'nanumsquare';
-    background-color: white;
-    position: fixed;
-    top: 0;
-    left: 0px;
-    width: 100%;
-    height: 100%;
-    display: flex;
-    justify-content:center;
-    align-items: center;
-    text-align: center;
-}
-
-.menu{
-    ul{
-        list-style: none;
-        padding: 0;
-        margin: 0;
-    }
-
-    li{
-        margin: 1rem 0;
-    }
-
-    a{
-        @include ani();
-        font-weight: 900;
-        font-size: 28px;
-        line-height: 80px;
-        padding-right: 30px;
-    }
-    a:hover{
-        opacity: 1;
-    }   
-}
-
-.xicon:hover{
-    opacity: 1;
-
-}
-.xicon{
-    position: relative;
-    bottom: 387px;
-    left: 238px;
-    font-size: 30px;
-    @include ani()
-}
